requirements.txt
et-xmlfile==1.1.0
numpy==1.23.3
openpyxl==3.0.10
Pillow==9.2.0
pip==22.2.2
python-dateutil==2.8.2
pytz==2022.2.1
qrcode==7.3.1
setuptools==65.3.0
six==1.16.0
wheel==0.37.1
